+++ packages/webapp/src/navbar/Pages.ts
@@ -9,6 +9,6 @@
  * Customize pages
  */
 export interface Pages {
-  disabled?: (("edit" | "date" | "video" | "map" | "tag") | string)[];
+  disabled?: (("edit" | "date" | "video" | "map" | "tag" | "folder") | string)[];
   [k: string]: unknown;
 }